refactor(types): convert to types.ts
