+++ packages/client-utils/src/types-index.d.ts
@@ -1,0 +1,2 @@
+// Export all the types this package provides
+export type * from './types.js';
